initial upload

+++ ArmPkg/Library/StandaloneMmMmuLib/ArmMmuStandaloneMmLib.c
@@ -68,13 +68,13 @@ SendMemoryPermissionRequest (
       // Now check the payload for errors
       // The callee sends back the return value
       // in Arg3
-      *RetVal = SvcArgs->Arg3;
+      *RetVal = (INT32)SvcArgs->Arg3;
     } else {
       // If Arg0 is not a Direct Response, that means we
       // have an FF-A error. We need to check Arg2 for the
       // FF-A error code.
       // See [2], Table 10.8: FFA_ERROR encoding.
-      *RetVal = SvcArgs->Arg2;
+      *RetVal = (INT32)SvcArgs->Arg2;
       switch (*RetVal) {
         case ARM_FFA_SPM_RET_INVALID_PARAMETERS:
           return EFI_INVALID_PARAMETER;
@@ -98,7 +98,7 @@ SendMemoryPermissionRequest (
       }
     }
   } else {
-    *RetVal = SvcArgs->Arg0;
+    *RetVal = (INT32)SvcArgs->Arg0;
   }
 
   // Check error response from Callee.
